requirements.txt
requests
numpy
pillow
rasterio
matplotlib
black
gdal
python-dotenv